pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-input"
version = "0.1.0"
description = "Voice input tool with multiple speech recognition backends"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "lifuhaolife"}
]
keywords = ["voice", "speech", "input", "whisper", "linux"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
]

dependencies = [
    "pynput>=1.7.6",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "openai-whisper>=20231117",
    "pyperclip>=1.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
voice-input = "voice_input.main:main"

[project.urls]
Homepage = "https://github.com/lifuhaolife/voice-input"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]